requirements.txt
Django==2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
psycopg2==2.8.6
Pillow==7.2.0
django-filter==2.2.0
django-environ==0.4.5